fix: qr download
